refactor(web): split dashboard.js into entry + view layer (v0.5.3.0)
dashboard.js had grown to 1023 lines mixing data fetching, model
derivation, HTML composition, and post-render side effects. Pragmatic
split:

  dashboard.js (483 lines)
    - Time/behavior constants + slowFetch + invalidateDashboardCache
    - renderDashboard: data → state → HTML composition → effects
    - Re-exports initDashboardGlobals so app.js's import is unchanged

  dashboard-view.js (NEW, 583 lines)
    - Label helpers: situationLabel, domainIcon, domainLabel,
      traitLabel, traitIcon, formatTime
    - Pure render helpers: renderNotificationOptIn, renderBriefingCard,
      renderSinceLastVisit, renderEmptyDashboardPreview,
      renderAskTwinWidget, renderAskResult, renderTourBanner,
      renderJustConnectedCelebration, renderConnectGoogleHero,
      renderUnmetCredentials, renderSkillGaps, renderBriefingItem
    - Global handlers: handleEnableNotifications, dismissNotifOptIn,
      handleAskTwin, skyTwinExitTour, handleConnectGoogleFromDashboard
    - initDashboardGlobals — wires the handlers + the document-level
      .ask-twin-example click delegator into window once

Zero runtime behavior change. Same imports, same window.X handlers,
same SSE events, same render output. The TODO had asked for a
three-way compute/view/effects split with testable model derivation —
deferred for a future pass when there's a concrete testing reason.
Today's win is the cognitive-load reduction.

Tests: pnpm test green across 38 packages.
