+++ TODOS.md
@@ -7,7 +7,7 @@ Generated from CEO review on 2026-04-01. Updated through M2/M3/M4 completion.
 Most items closed in the same branch (see Completed below). Three items
 remain open as P2/P3 follow-ups.
 
-- [ ] **P2**: Split `renderDashboard` (~340 lines now) into `computeDashboardModel`, `renderDashboardView`, `applyDashboardEffects`. The current function mixes data fetching, state derivation, HTML composition, and post-render side effects — hard to test, hard to refactor. **Files:** apps/web/public/js/pages/dashboard.js
+- [x] **P2**: ~~Split `renderDashboard`~~ **Completed:** v0.5.3.0 (2026-04-29) — Pragmatic split: presentation moved to `dashboard-view.js` (pure render helpers + handlers + `initDashboardGlobals`); `dashboard.js` shrinks from 1023 → 483 lines focused on data flow + lifecycle. Full compute/view/effects extraction left for a future pass when there's a concrete testing reason; today's win is the cognitive-load reduction. Zero runtime behavior change.
 - [ ] **P2**: Migrate the remaining inline `onclick="...handleX('${escapeHtml(value)}')"` patterns across approvals.js / settings.js / decisions.js / setup.js / dashboard.js to `addEventListener` bindings. The highest-risk site (twin.js, free-text preference value flowing through JS string literals) was migrated in this PR. The remaining ~40 inline handlers interpolate UUIDs / enum values / constants where practical risk is zero today, but the pattern is unsafe by construction. **Files:** apps/web/public/js/pages/{approvals,settings,decisions,setup,dashboard}.js
 - [x] **P3**: ~~OAuth callback redirect contract change verification~~ **Completed:** 2026-04-29 — Audited `apps/desktop/`, `apps/mobile/`, README, all `*.ts/*.tsx/*.js/*.md` repo-wide. No consumers reference the old `#/settings?connected=…` shape. Desktop OAuth uses an HTML success page (no client-side route consumed). Mobile sse-client has unrelated `callback` mentions. Safe to ship the new `#/?connected=…` redirect without backwards-compat. **Verified by:** grep audit on origin/main HEAD `a72e85b`.
 - [ ] **P3**: Real production tour mode — short-lived demo session token, or `/api/v1/demo/dashboard` aggregate read-only endpoint. Currently tour mode auto-disables in production (auth bypass not active), which is honest but cuts off the marquee feature for non-localhost deploys. **Files:** apps/api/src/routes/demo.ts, apps/api/src/middleware/session-auth.ts
