+++ VERSION
@@ -1,1 +1,1 @@
-0.5.2.0
+0.5.3.0
