+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "skytwin",
-  "version": "0.5.2.0",
+  "version": "0.5.3.0",
   "description": "A delegated judgment layer that builds a digital twin of your preferences and acts on your behalf — with safety constraints, explanations, and progressive trust.",
   "private": true,
   "repository": {
